Console Commands
You can now define a selection of commands to run on the console in a commands.lst file. These will be placed in the Console Commands menu.
This is particularly useful if you a) just want to cheat or b) need to skip a scene or to a particular map because a mod isn't working correctly.